requirements.txt
Flask==3.0.0
Gunicorn==21.2.0
aiohttp==3.9.1
urllib3==2.1.0
pycryptodome==3.19.0
protobuf-decoder==0.2.0
